Reorganise components
